Refactor common formatters to be common helper
