+++ test/acceptance/features/location/step_definitions/location.js
@@ -1,26 +1,8 @@
 const { find, assign, set, get, camelCase, includes } = require('lodash')
 const { client } = require('nightwatch-cucumber')
 const { defineSupportCode } = require('cucumber')
-const { addMinutes, isAfter } = require('date-fns')
-const moment = require('moment')
 
-const { mediumDateTimeFormat } = require('../../../../../config')
-
-const formatters = {
-  isEuropeanOrGlobalHeadquartersFormatter (expected, actual) {
-    const formatted = /^(european|global) headquarters$/i.test(expected) ? 'Yes' : 'No'
-    return formatted === actual
-  },
-  isProjectCodeFormatter (expected, actual) {
-    return /^DHP-[0-9]{8}$/i.test(actual)
-  },
-  isRecentDateFormatter (expected, actual) {
-    const actualDate = moment(actual, mediumDateTimeFormat).toDate()
-    const oneMinuteAgo = addMinutes(actualDate, -1)
-
-    return isAfter(actualDate, oneMinuteAgo)
-  },
-}
+const formatters = require('../../../helpers/formatters')
 
 function getExpectedValue (row, state) {
   if (includes(row.value, '.') && !includes(row.value, ' ')) {
